tracing-demo-resources/requirements.txt
blinker==1.8.2
click==8.1.7
Deprecated==1.2.14
flask==3.0.3
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
opentelemetry-api==1.26.0
opentelemetry-distro==0.47b0
opentelemetry-instrumentation==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
typing-extensions==4.12.2
werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.0
